+++ drivers/clock_control/CMakeLists.txt
@@ -8,14 +8,14 @@ zephyr_sources_ifdef(CONFIG_CLOCK_CONTROL_QUARK_SE            quark_se_clock_con
 zephyr_sources_ifdef(CONFIG_CLOCK_CONTROL_RV32M1_PCC          clock_control_rv32m1_pcc.c)
 
 if(CONFIG_CLOCK_CONTROL_STM32_CUBE)
-  zephyr_sources(stm32_ll_clock.c)
+  zephyr_sources(clock_stm32_ll_common.c)
 
-  zephyr_sources_ifdef(CONFIG_SOC_SERIES_STM32F0X stm32f0x_ll_clock.c)
-  zephyr_sources_ifdef(CONFIG_SOC_SERIES_STM32F1X stm32f1x_ll_clock.c)
-  zephyr_sources_ifdef(CONFIG_SOC_SERIES_STM32F2X stm32f2x_ll_clock.c)
-  zephyr_sources_ifdef(CONFIG_SOC_SERIES_STM32F3X stm32f3x_ll_clock.c)
-  zephyr_sources_ifdef(CONFIG_SOC_SERIES_STM32F4X stm32f4x_ll_clock.c)
-  zephyr_sources_ifdef(CONFIG_SOC_SERIES_STM32F7X stm32f7x_ll_clock.c)
-  zephyr_sources_ifdef(CONFIG_SOC_SERIES_STM32L0X stm32l0x_ll_clock.c)
-  zephyr_sources_ifdef(CONFIG_SOC_SERIES_STM32L4X stm32l4x_ll_clock.c)
+  zephyr_sources_ifdef(CONFIG_SOC_SERIES_STM32F0X clock_stm32f0x.c)
+  zephyr_sources_ifdef(CONFIG_SOC_SERIES_STM32F1X clock_stm32f1x.c)
+  zephyr_sources_ifdef(CONFIG_SOC_SERIES_STM32F2X clock_stm32f2x.c)
+  zephyr_sources_ifdef(CONFIG_SOC_SERIES_STM32F3X clock_stm32f3x.c)
+  zephyr_sources_ifdef(CONFIG_SOC_SERIES_STM32F4X clock_stm32f4x.c)
+  zephyr_sources_ifdef(CONFIG_SOC_SERIES_STM32F7X clock_stm32f7x.c)
+  zephyr_sources_ifdef(CONFIG_SOC_SERIES_STM32L0X clock_stm32l0x.c)
+  zephyr_sources_ifdef(CONFIG_SOC_SERIES_STM32L4X clock_stm32l4x.c)
 endif()
