+++ drivers/clock_control/clock_stm32l4x.c
@@ -1,6 +1,6 @@
 /*
  *
- * Copyright (c) 2018 qianfan Zhao
+ * Copyright (c) 2017 Linaro Limited.
  *
  * SPDX-License-Identifier: Apache-2.0
  */
@@ -11,7 +11,7 @@
 #include <clock_control.h>
 #include <misc/util.h>
 #include <clock_control/stm32_clock_control.h>
-#include "stm32_ll_clock.h"
+#include "clock_stm32_ll_common.h"
 
 
 #ifdef CONFIG_CLOCK_STM32_SYSCLK_SRC_PLL
@@ -20,8 +20,8 @@
 #define z_pllm(v) LL_RCC_PLLM_DIV_ ## v
 #define pllm(v) z_pllm(v)
 
-#define z_pllp(v) LL_RCC_PLLP_DIV_ ## v
-#define pllp(v) z_pllp(v)
+#define z_pllr(v) LL_RCC_PLLR_DIV_ ## v
+#define pllr(v) z_pllr(v)
 
 /**
  * @brief fill in pll configuration structure
@@ -30,7 +30,7 @@ void config_pll_init(LL_UTILS_PLLInitTypeDef *pllinit)
 {
 	pllinit->PLLM = pllm(CONFIG_CLOCK_STM32_PLL_M_DIVISOR);
 	pllinit->PLLN = CONFIG_CLOCK_STM32_PLL_N_MULTIPLIER;
-	pllinit->PLLP = pllp(CONFIG_CLOCK_STM32_PLL_P_DIVISOR);
+	pllinit->PLLR = pllr(CONFIG_CLOCK_STM32_PLL_R_DIVISOR);
 }
 #endif /* CONFIG_CLOCK_STM32_SYSCLK_SRC_PLL */
 
@@ -39,14 +39,24 @@ void config_pll_init(LL_UTILS_PLLInitTypeDef *pllinit)
  */
 void config_enable_default_clocks(void)
 {
-	/* Power Interface clock enabled by default */
-	LL_APB1_GRP1_EnableClock(LL_APB1_GRP1_PERIPH_PWR);
-}
+#ifdef CONFIG_CLOCK_STM32_LSE
+	/* LSE belongs to the back-up domain, enable access.*/
 
-/**
- * @brief Function kept for driver genericity
- */
-void LL_RCC_MSI_Disable(void)
-{
-	/* Do nothing */
+	/* Enable the power interface clock */
+	LL_APB1_GRP1_EnableClock(LL_APB1_GRP1_PERIPH_PWR);
+
+	/* Set the DBP bit in the Power control register 1 (PWR_CR1) */
+	LL_PWR_EnableBkUpAccess();
+	while (!LL_PWR_IsEnabledBkUpAccess()) {
+		/* Wait for Backup domain access */
+	}
+
+	/* Enable LSE Oscillator (32.768 kHz) */
+	LL_RCC_LSE_Enable();
+	while (!LL_RCC_LSE_IsReady()) {
+		/* Wait for LSE ready */
+	}
+
+	LL_PWR_DisableBkUpAccess();
+#endif
 }
